+++ algo/heuristic_test.go
@@ -6,15 +6,15 @@ import (
 )
 
 func TestManhattan(t *testing.T) {
-	var res []int8
-	var dist int8
-	a := []int8{3, 0, 5, 8, 1, 4, 2, 6, 7}
+	var res []int16
+	var dist int16
+	a := []int16{3, 0, 5, 8, 1, 4, 2, 6, 7}
 	res = checker.BuildCorrectResult(3)
 	dist = CalculateManhattanDistance(a, res)
 	if dist != 11 {
 		t.Errorf("Error: Manhattan Distance should be 11 and not %d", dist)
 	}
-	b := []int8{8, 2, 7, 4, 1, 0, 10, 15, 3, 9, 6, 13, 11, 5, 12, 14}
+	b := []int16{8, 2, 7, 4, 1, 0, 10, 15, 3, 9, 6, 13, 11, 5, 12, 14}
 	res = checker.BuildCorrectResult(4)
 	dist = CalculateManhattanDistance(b, res)
 	if dist != 37 {
@@ -23,8 +23,8 @@ func TestManhattan(t *testing.T) {
 }
 
 func TestGetIndexOf(t *testing.T) {
-	a := []int8{3, 5, 1, 8, 0, 4, 2, 6, 7}
-	var b int8 = 4
+	a := []int16{3, 5, 1, 8, 0, 4, 2, 6, 7}
+	var b int16 = 4
 	index := GetIndexOf(b, a)
 	if index != 5 {
 		t.Errorf("Error: index should be 5 and not %d", index)
